feat(Commands): Opening GUIs for another player is now allowed
